updated save method
