+++ server/charts/server/templates/deployment.yaml
